pyproject.toml
[tool.poetry]
name = "zkteco-to-db"
version = "0.1.0"
description = ""
authors = ["Thanaphat T. <ice48623@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.*"
pyzk = "^0.9"
peewee = "^3.17.1"
psycopg2-binary = "^2.9.9"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
xlsxwriter = "^3.2.2"
sendgrid = "^6.12.2"
brevo-python = "^1.2.0"
pyyaml = "^6.0.2"


[tool.poetry.group.dev.dependencies]
setuptools = "^80.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
